Refactor all hand mesh rendering into a new HandMeshRenderer class hierarchy
This brings sanity and encapsulation to existing hand mesh rendering
code, which has been evolving rather experimentally so far. It also
prepares for the introduction of yet another mesh rendering mechanism
(XR_MSFT_hand_tracking_mesh) in a follow-up series.

Currently there are two rendering strategies: `Spheres` (Pico and Spaces),
and `Skinned` (all other devices). A new strategy `Geometry` will be
added for the MSFT extension later.

The refactoring introduces a new base class, HandMeshRenderer, which has
three basic methods: Initialize(), Update() and Draw(). Derived classes
will implement one rendering strategy each. With only these 3 methods,
we are able to accommodate all 3 rendering strategies in a simple and
coherent manner.

There is one single `handMeshRenderer` object that belongs to
BrowserWorld. It is created lazily upon having to draw() hands (meaning
hand tracking is already active for at least one hand).

The Initialize() method is called automatically when the instance is
created.
Update() is called every time controllers state is updated via
BrowserWorld::UpdateControllers().
Finally, Draw() is called on BrowserWorld::DrawWorld(), after all other
elements has been drawn (except the controller models).

The old class HandModels that was used for Skinned mesh rendering is
completely replaced by HandMeshRendererSkinned, and thus its files
are removed.

The patch has been tested on Quest2 and Lenovo A3.
